more changes

+++ cdf/lib/importable.rb
@@ -168,39 +168,45 @@ module Importable
     end
 
     def download_file(client, file_name)
+      # client = client || default_client
       client = CdfFtpClient.new({:keep_alive => true, :server => ftp_server, :user => ftp_user_name, :password => ftp_password }) if client.nil?
       remote_listing = client.dir ftp_dirs.first, ".*#{@ext}"
       import_file = self.new_or_archived(file_name)
 
       if zip_file?
+        data_file_name = file_name
         file_name = file_name.partition(".")[0] + ".#{@ext}"
       end
+
       local_path = create_path file_name
       client.get file_name, local_path
 
       if zip_file?
-        Zip::ZipFile.open(local_path) { |zip_file|
-          f_path=File.join(CdfConfig::current_data_lib_in, zip_file.first.name)
-          if File.exists? f_path
-            File.delete f_path
-          end
-          zip_file.extract(zip_file.first, f_path)          # extract from the zip file
+        f_path=File.join(CdfConfig::current_data_lib_in, data_file_name)
+        if File.exists? f_path
+          File.delete f_path
+        end
 
-          # split up the extracted files into more manageable parts
-          prefix = import_file.generate_part_file_prefix
-          Dir.foreach(CdfConfig::current_data_lib_in) do |part_file|
-            File.delete File.join(CdfConfig::current_data_lib_in, part_file) if part_file.starts_with? prefix    # delete any old parts of this file, if any are present
-          end
-          system("cd #{File.dirname(f_path)} && split -a 4 #{f_path} #{prefix}")    # split up the new file into portions
-          Dir.foreach(CdfConfig::current_data_lib_in) do |part_file|
-            if part_file.starts_with? prefix
-              write_data_with_delimiters File.join(CdfConfig::current_data_lib_in, part_file)
-            end
-          end
+        system("cd #{File.dirname(f_path)} && unzip #{file_name} | xargs chmod 666")
+        if !File.exists? f_path
+          raise "Unexpected result after unzipping the downloaded file!"
+        end
 
-          import_file.file_name = "#{zip_file.first.name}"
-          import_file.save
-        }
+        # split up the extracted files into more manageable parts
+        prefix = import_file.generate_part_file_prefix
+        Dir.foreach(CdfConfig::current_data_lib_in) do |part_file|
+          File.delete File.join(CdfConfig::current_data_lib_in, part_file) if part_file.starts_with? prefix    # delete any old parts of this file, if any are present
+        end
+
+        system("cd #{File.dirname(f_path)} && split -l 1000 -a 4 #{f_path} #{prefix}")    # split up the new file into portions
+        Dir.foreach(CdfConfig::current_data_lib_in) do |part_file|
+          if part_file.starts_with? prefix
+            write_data_with_delimiters File.join(CdfConfig::current_data_lib_in, part_file)
+          end
+        end
+
+        import_file.file_name = data_file_name
+        import_file.save
       else
         write_data_with_delimiters local_path
       end
@@ -259,6 +265,10 @@ module Importable
     def create_path(file_name)
       File.join CdfConfig::current_data_lib_in, file_name
     end
+
+#    def copy_file src, dest
+#      system("cp #{src} #{dest}")
+#    end
 
     def needs_import
       where("#{self.table_name}.imported_at IS NULL")
